+++ Classes/KMLLocation.m
@@ -69,7 +69,7 @@
 
 - (void)setAltitude:(NSInteger)altitude
 {
-    _altitudeValue = [NSString stringWithFormat:@"%d", altitude];
+	_altitudeValue = [NSString stringWithFormat:@"%ld", (long)altitude];
 }
